Backend/requirements.txt
Django>=5.0,<5.1
psycopg2-binary>=2.9
python-dotenv>=1.0
djangorestframework>=3.15
django-filter>=24.2
Pillow>=10.4